BUG/MINOR: cli: Report an error to user if command or payload is too big
Too big command, larger than a buffer, was silently rejected by the CLI
applet. It was handled as an error and the connection was closed, but no
error message was reported to user to notify him. Now an error is reported
before closing. It is only displayed if the chunk buffer used by the CLI
applet is full and no delimiter (\n or ;) is found to mark the end of the
command. It works for a simple command but also for a command with a huge
payload.

This patch could be backported to all stable versions.

(cherry picked from commit 498520fdf505f05cde2f99c70c4e98fe90098961)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit a67b18d658e153c2d8c55763946d3c62f8ba8a36)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>