BUG/MINOR: server: ignore 'enabled' for dynamic servers
This commit is an adaptation from the following patch introduced in 3.0
dev tree.

  250c19032f030a81dca01541e3c2947d117a041c
  BUG/MINOR: server: reject enabled for dynamic server

Instead of reporting an error on 'add server', 'enabled' keyword is
simply ignored. This guarantees that CLI API remains compatible for
stable versions. A warning has been added to signal this status before
the final message.

This should be backported up to 2.4. The warning should be removed on
2.7 and below versions to prevent error message flooding for users
stucked on older versions.