BUG/MINOR: http-ana: Fix TX_L7_RETRY and TX_D_L7_RETRY values
These values are obviously wrong. There is an extra zero at the end for both
defines. By chance, it is harmless. But it is better to fix it.

This patch should be backported as far as 2.6.

(cherry picked from commit 1fa6eb2eb9fa3300d564faae7d1a6d4e7ad49389)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>