BUG/MEDIUM: spoe: Always retry when an applet fails to send a frame
This bug is related to the previous one ("BUG/MEDIUM: spoe: Always retry
when an applet fails to send a frame"). applet_putblk() function retruns -1
on error and it should always be interpreted as a missing of room in the
buffer. However, on the spoe, this was processed as an I/O error.

This patch must be backported as far as 2.8.

(cherry picked from commit 4fd656e311d7817e03e0f9ddca4c6635ccd4476f)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>